exp2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
einops>=0.6.0
tqdm>=4.64.0
numpy>=1.21.0
Pillow>=9.0.0